+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- [#21](https://github.com/increments/qiita_marker/pull/21): Add `AUTOLINK_CLASS_NAME` option for autolink extension.
+
 ### Changed
 
 - [#19](https://github.com/increments/qiita_marker/pull/19): Allow a custom block to contain any node type.